Added and tested all CRUD operations for workout plans. Made changes in user and workout plan models

+++ controllers/workoutPlanController.js
@@ -8,10 +8,11 @@ const workoutPlanController = {
       const workoutPlans = await workoutPlanModel
         .find({})
         .populate("workouts.workoutId")
-        .populate("created_by");
-      res.json(workoutPlans);
+        .populate("created_by")
+        .sort({ createdAt: -1 });
+      return res.status(200).json(workoutPlans);
     } catch (error) {
-      res.status(500).json({ message: error.message });
+      return res.status(500).json({ error: error.message });
     }
   },
 
@@ -23,13 +24,14 @@ const workoutPlanController = {
       }
       const workoutPlan = await workoutPlanModel
         .findById(req.params.id)
-        .populate("workouts.workoutId");
+        .populate("workouts.workoutId")
+        .populate("created_by");
       if (!workoutPlan) {
         return res.status(404).json({ message: "Workout plan not found" });
       }
-      res.json(workoutPlan);
+      return res.status(200).json(workoutPlan);
     } catch (error) {
-      res.status(500).json({ message: error.message });
+      return res.status(500).json({ error: error.message });
     }
   },
 
@@ -38,18 +40,22 @@ const workoutPlanController = {
     try {
       const workoutPlan = new workoutPlanModel({
         ...req.body,
-        createdBy: req.user._id, // Assuming the user's ID is available from the request
+        // createdBy: req.user._id, // Assuming the user's ID is available from the request
       });
       await workoutPlan.save();
-      res.status(201).json(workoutPlan);
+      return res.status(200).json(workoutPlan);
     } catch (error) {
-      res.status(400).json({ message: error.message });
+      return res.status(400).json({ error: error.message });
     }
   },
 
   // Update a workout plan
   updateWorkoutPlan: async (req, res) => {
     try {
+      // checks if id parameter is valid or not
+      if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return res.status(404).json({ error: "Workout plan not found" });
+      }
       const updatedWorkoutPlan = await workoutPlanModel.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -58,24 +64,28 @@ const workoutPlanController = {
       if (!updatedWorkoutPlan) {
         return res.status(404).json({ message: "Workout plan not found" });
       }
-      res.json(updatedWorkoutPlan);
+      return res.status(200).json(updatedWorkoutPlan);
     } catch (error) {
-      res.status(400).json({ message: error.message });
+      return res.status(400).json({ error: error.message });
     }
   },
 
   // Delete a workout plan
   deleteWorkoutPlan: async (req, res) => {
     try {
+      // checks if id parameter is valid or not
+      if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return res.status(404).json({ error: "Workout plan not found" });
+      }
       const workoutPlan = await workoutPlanModel.findByIdAndDelete(
         req.params.id
       );
       if (!workoutPlan) {
         return res.status(404).json({ message: "Workout plan not found" });
       }
-      res.json({ message: "Workout plan successfully deleted" });
+      return res.status(200).json(workoutPlan);
     } catch (error) {
-      res.status(500).json({ message: error.message });
+      return res.status(500).json({ error: error.message });
     }
   },
 };
